Revert "POL-119 Verify from RBAC each user account and if they access to policies before sending the mail"
This reverts commit 769dce6f9759f38793d3f1f4a64218f04629d8df.
