+++ app/core/errors.py
@@ -17,7 +17,3 @@ class BOPInvalidRecipientException(BOPException):
 
 class NoTemplateFoundException(Exception):
     pass
-
-
-class RbacException(Exception):
-    pass
